Crate os_type

Structs

OSInformation
Holds information about Operating System type and its version If the version could not be fetched it defaults to 0.0.0

Enums

OSType
A list of supported operating system types

Functions

current_platform
Returns the current operating system type